cleaning up label field filters
